+++ count/hll.cc
@@ -137,7 +137,7 @@ double HLL::RawEstimate() const {
   return estimate;
 }
 
-uint64_t HLL::Estimate() const {
+double HLL::Estimate() const {
   // TODO(tdial): The logic below was more or less copied from the research
   // paper, less the handling of the sparse register array, which is not
   // implemented at this time. It is correct, but seems a little awkward.
